New endpoint /upload_file_size for smaht-submitr usage.
