+++ CHANGELOG.rst
@@ -19,11 +19,12 @@ Change Log
 * Added scripts/opensearch-dashboard-start.bash and Makefile target opensearch-dashboard-starto
   now that we are using OpenSearch rather than ElasticSearch we seem to need this rather than Kibana.
 * Added demo directory with script to export some sample/demo from smaht-staging and import locally.
-* Added /upload_file_exists/{upload_file_uuid}/{upload_file_name} HEAD endpoint in metadata.py; for
-  use by smaht-submitr to determine if a file to upload has already been uploaded; and in particular
-  if it has been uploaded but is still marked as 'uploading' because its md5 is still in the process
-  of being computed (i.e. otherwise we could simply look at the file status an be done with it).
-  Also added similar /update_file_size GET endpoint to get this uploaded file size if exists.
+* Added /files/upload_file_size endpoint (types/file.py); for use by smaht-submitr to determine if a file to
+  upload has already been uploaded; and get its size as a side-effect; returns HTTP 200 if found otherwise 404);
+  in particular we want to check if it has been uploaded but is still marked as 'uploading' because its md5 is
+  still in the process of being computed (i.e. otherwise we could simply look at the file status an be done with it).
+* New protection of /ingestion-status endpoint (in ingestion/ingestion_status.py) for authenticated,
+  i.e. logged in, users only. Also protect/limit /ingestion-status/{keys,keys_sorted,flush} to admin.
 
 
 0.85.0
