+++ src/encoded/types/file.py
@@ -1,16 +1,18 @@
-import functools
+from boto3 import client as boto_client
 from datetime import datetime
+import functools
+from pyramid.exceptions import HTTPForbidden
+from pyramid.request import Request
+from pyramid.response import Response
+from pyramid.view import view_config
 from typing import Any, Dict, List, Optional, Union
 
-from pyramid.view import view_config
 from encoded_core.types.file import (
     HREF_SCHEMA,
     UNMAPPED_OBJECT_SCHEMA,
     UPLOAD_KEY_SCHEMA,
     File as CoreFile,
 )
-from pyramid.request import Request
-from pyramid.exceptions import HTTPForbidden
 from encoded_core.file_views import (
     validate_file_filename,
     validate_extra_file_format,
@@ -996,6 +998,31 @@ def download(context, request):
         raise HTTPForbidden('This is a restricted file not available for download without dbGAP approval. '
                             'Please check with DAC/your PI about your status.')
     return CoreDownload(context, request)
+
+
+# This /files/upload_file_size endpoint was added specifically (2024-08-22) for smaht-submitr to
+# determine if a file to upload has already been uploaded, and to get its size as a side-effect,
+# to report to the submitter; if the file does not exist then HTTP 404 is returned, otherwise
+# HTTP 200 is returned and also its (byte) size, in the "size" element of the returned JSON.
+@view_config(name='upload_file_size', context=File, permission='view', request_method=['GET'])
+@debug_log
+def upload_file_size(context, request):
+    if upload_key := context.upload_key(request):
+        if not (upload_bucket := request.GET.get("upload_bucket")):
+            upload_bucket = request.registry.settings.get("file_upload_bucket")
+        if upload_bucket:
+            s3 = boto_client("s3")
+            try:
+                response = s3.head_object(Bucket=upload_bucket, Key=upload_key)
+                if isinstance(file_size := response.get("ContentLength", None), int):
+                    return {
+                        "bucket": upload_bucket,
+                        "key": upload_key,
+                        "size": file_size
+                    }
+            except Exception:
+                pass
+    return Response(status=404)
 
 
 def validate_processed_file_unique_md5_with_bypass(context, request):
